Extract load_json and dump_json in a module and reuse in tests
