+++ generate_metadata_schema_files.py
@@ -1,18 +1,9 @@
 import argparse
 import json
 import os
+import file_util
 
 from metadata_schema import MetadataSchema
-
-
-def load_json(file_path):
-    f = open(file_path, 'r')
-    return json.loads(f.read())
-
-
-def dump_json(file_path, data):
-    with open(file_path, 'w') as outfile:
-        json.dump(data, outfile, indent=4, sort_keys=True)
 
 
 def get_schema_paths(base_path):
